switched handler and producer fixed info print in main
