+++ templates/src/handler/$$handler$$.rs.go
@@ -1,62 +1,81 @@
 use async_nats::{Client, Message, jetstream};
 use async_nats::jetstream::Context;
-use crate::{publish_message,stream_publish_message,model::*,config::*};
+use crate::{publish_message,stream_publish_message,model::*,config::*,policy::*};
 use std::time;
 use opentelemetry::global;
 use opentelemetry::trace::Tracer;
 use log::{debug, warn};
 
     {{ $isStream := false }}
-    {{ $channel := . }}
 
-    /// Publish a message in the {{ .unique_id }} channel
+    /// This handler is called when a message is received on channel {{ .unique_id }}
     /// Channel messages:
     /// {{ range .messages }}
     ///     {{ .unique_id }}
     /// {{ end }}
-    {{ if key_exists  "original_operation" "bindings" "nats" "streamname" }}
-        {{ $isStream := .original_operation.bindings.nats.streamname }}
-    {{ end }}
-            
-    {{ if $isStream }}
+    {{if key_exists  "original_operation" "bindings" "nats" "streamname"}}
+        {{ $isStream := (.original_operation.bindings.nats.streamname) }}
+    {{end}}
+    {{if $isStream}}
+        pub fn stream_handler_{{ .unique_id }}(message: jetstream::Message, client: &Client) {
+        let tracer = global::tracer("stream_handler_{{ .unique_id }}");
+        let _span = tracer.start("{{ .unique_id }}_stream_handler");
         {{ range .messages }}
-            pub async fn stream_producer_{{ .unique_id }}(context_stream: &Context, payload : {{ if .payload}} {{ .payload.struct_reference }} {{ else }} () {{ end }}) { //context instead of client
-    let tracer = global::tracer("{{ .unique_id }}_stream_producer");
-        let _span = tracer.start("stream_producer_{{ .unique_id }}");
-        let subject = get_env().get("{{ .unique_id }}_SUBJECT").unwrap().clone();
-                {{ if .payload }}
-                    let payload = match serde_json::to_string(&payload) {
-                        Ok(payload) => payload,
+                {{ if .payload}}
+                    match serde_json::from_slice::<{{ .payload.struct_reference }}>(&message.message.payload.as_ref()) {
+                        Ok(deserialized_message) => {
+                            debug!("Received message {:#?}", deserialized_message);
+                            let policy_reply = opa_eval(message);
+                            // TODO: Replace this with your own handler code
+                            {{ if eq .payload.model_type "enum"}}
+                                match deserialized_message {
+                                    {{$enumName := .payload.unique_id}}
+                                    {{ range .payload.related_models }}
+                                        {{ $enumName }}::{{ .unique_id }}(payload) => {
+                                        // TODO: Replace this with your own handler code
+                                        println!("Received message payload {{ .unique_id }} {:?}", payload);
+                                        }
+                                    {{ end }}
+                                }
+                            {{ end }}
+                        },
                         Err(_) => {
-                            warn!("Failed to serialize message payload: {{ .payload.struct_reference }}");
-                            return;
-                        }
-                    };
-                    stream_publish_message(context_stream, &subject, &payload).await;
-                {{else}}
-                stream_publish_message(context_stream, &subject, &"").await;
-                {{end}}
-            }
-        {{end}}
-    {{ else }}
+                            warn!("Failed to deserialize message payload: {{ .unique_id }}\nOriginal message: {:#?}", message);
+                            // TODO: Handle the failed deserialization here
+                        },
+                    }
+                {{ end }}
+            {{ end }}
+        }
+    {{else}}
+        pub async fn handler_{{ .unique_id }}(message: Message, client: &Client) {
+    let tracer = global::tracer("handler_{{ .unique_id }}");
+            let _span = tracer.start("{{ .unique_id }}_handler");
     {{ range .messages }}
-    pub async fn producer_{{ .unique_id }}(client: &Client, payload: {{ if .payload }} {{.payload.struct_reference}} {{else}} () {{end}}) {
-    let tracer = global::tracer("{{ .unique_id }}_producer");
-    let _span = tracer.start("producer_{{ .unique_id }}");
-    let subject = get_env().get("{{ .unique_id }}_SUBJECT").unwrap().clone();
-                {{ if .payload }}
-                    let payload = match serde_json::to_string(&payload) {
-                        Ok(payload) => payload,
-                        Err(_) => {
-                            warn!("Failed to serialize message payload: {{ .payload.struct_reference }}");
-                            return;
-                        }
-                    };
-                    publish_message(client, &subject, &payload).await;
-                {{else}}
-                    publish_message(client, &subject, &"").await;
-                {{end}}
-            }
-        {{ end }}
-    {{ end }}
-
+                {{ if .payload}}
+                match serde_json::from_slice::<{{ .payload.struct_reference }}>(&message.payload.as_ref()) {
+                    Ok(deserialized_message) => {
+                        {{ if eq .payload.model_type "enum"}}
+                            match deserialized_message {
+                                {{$enumName := .payload.unique_id}}
+                                {{ range .payload.related_models }}
+                                    {{ $enumName }}::{{ .unique_id }}(payload) => {
+                                    // TODO: Replace this with your own handler code
+                                    debug!("Received message payload {{ .unique_id }} {:?}", payload);
+                                    }
+                                {{ end }}
+                            }
+                        {{else}}
+                            debug!("Received message {:#?}", deserialized_message);
+                            // TODO: Replace this with your own handler code
+                        {{ end }}
+                    },
+                    Err(_) => {
+                        warn!("Failed to deserialize message payload: {{ .unique_id }}\nOriginal message: {:#?}", message);
+                        // TODO: Handle the failed deserialization here
+                    },
+                }
+                {{ end }}
+            {{ end }}
+        }
+    {{end}}
